Feat: 파일 구조 정리 및 Header Component 작성
