+++ perfume-website/src/App.js
@@ -1,0 +1,12 @@
+import './App.css';
+import Header from './Components/Header'
+
+function App() {
+  return (
+    <div className="App">
+      <Header />
+    </div>
+  );
+}
+
+export default App;
